move graphql inside component
as per https://the-guild.dev/blog/unleash-the-power-of-fragments-with-graphql-codegen
