+++ codegen.ts
@@ -3,7 +3,7 @@ import type { CodegenConfig } from '@graphql-codegen/cli';
 const config: CodegenConfig = {
   overwrite: true,
   schema: 'http://localhost:4000/graphql',
-  documents: './lib/queries/*.js',
+  documents: ['./**/*.tsx', '!node_modules/**'],
   generates: {
     'src/gql/': {
       preset: 'client',
