+++ components/ProjectCard.tsx
@@ -1,12 +1,26 @@
-import { PropsWithChildren } from 'react';
 import { Text, StyleSheet } from 'react-native';
 import { FragmentType, useFragment } from '../src/gql/fragment-masking';
-import { ProjectFragment } from '../lib/queries/projects';
+import { graphql } from '../src/gql/gql';
 import color from '../styles/colors';
 
-const ProjectCard = (
-  props: PropsWithChildren<{ item: FragmentType<typeof ProjectFragment> }>
-) => {
+export const ProjectFragment = graphql(/* GraphQL */ `
+  fragment ProjectItem on Project {
+    _count {
+      tickets
+    }
+    createdAt
+    id
+    limitDate
+    name
+    progress
+  }
+`);
+
+type ProjectProps = {
+  item: FragmentType<typeof ProjectFragment>;
+};
+
+const ProjectCard = (props: ProjectProps) => {
   const projectData = useFragment(ProjectFragment, props.item);
   return <Text style={styles.project}>{projectData.name}</Text>;
 };
